Add more robust application loading/validation and improved database access
Added a separate function in loadApplications.ts to perform more extensive validation of an application before loading. This includes checking for existence of required files for each application: index.js, application.json and icon.avif and log errors when required files is missing. This change makes the code more maintainable and easier to understand.

Modified global_db/index.ts to allow for setting values in the global database via a POST request, improving the flexibility and functionality of database interaction.

Adjusted shutdown handling in main.ts to ensure the global database is written to disk before process termination.

Lastly, adjusted the path used for fetching data in globalDbApplication.tsx to match the changes made in global_db/index.ts. Changes were made to streamline the code and improve readability.