Update the app's internal structure and refactor the code
The changes include restructuring of the app which involved moving some files and removing others. This was done to better organize the code and makes it easier to manage as the codebase grows. Several files were also refactored to improve the readability and maintainability of the code. The token classification features were added to the code editor, and the import and export function feature was removed from the token.ts file for improved organization. Major changes also include the addition of the feature to switch between the 'browser' and 'dashboard' layouts in the dash application. Another major change is the implementation of localized greetings in the app, which is aimed at providing a better user experience. Changes were also made to the login Page for better user management facilities. The URLs were updated to include "/core" before "/login". Lastly, some enhancements were done in the PSA-backend to improve the performance of the app. The key aim of all these changes was to improve code organization, provide better UI experience and enhance app performance.