"Update code styling and add SVG icon mappings in project files"
This commit removes the CHAINED_CALL_DOT_ON_NEW_LINE option in the code styles settings and also adds a large list of SVG icon mappings in the 'icons.json' file. This was necessary to update our code styling rules and for planning to add more available icons for UI development.