Update Chip.module.scss and add new icon.svg
Icon.svg file was added to the project, this file contained an SVG image that is going to be used in the UI. Changes were made to the Chip.module.scss file for styling component that is of type "button". This was done to improve the look and feel of those buttons when they are rendered in the application.