Refactor codebase to use IconDictionary for icon references
The refactoring process included replacing explicit icon names-- passed as strings to the Icon component-- with references to an IconDictionary. This method ensures type safety and reduces chances of typing incorrect or non-existent icons. Moreover, this aids the future development efforts as the IDE would provide auto-suggestions to the dev while writing code. Other changes consist of improved code formatting, for instance, we have avoided an anomaly by using curly braces in a more consistent fashion. This commit doesn't impact any functionality, it is a part of codebase optimization to improve efficiency and maintainability.