Update layout, improve program functions, and fix style issues
Implemented a more robust and user-friendly layout for application cards and improved the transition animation when switching view states in the launcher. The personal server accelerator was also updated to include logs and maintain a history file, offering better debugging experiences. The user database was modified to provide more optimized searches and patterns. Improved the consistency of color coding for elements and fixed issues with incorrect CSS styling for buttons and borders. An issue with the CSV separator was resolved to not function properly. Lastly, fixed issues with incorrect version deployment and properly redirected to documentation pages in the application.